Bug Fixed

+++ RealBuildPOC.UserService/UserService.Application/Interfaces/IUserRepository.cs
@@ -4,8 +4,8 @@ namespace UserService.Application.Interfaces
 {
     public interface IUserRepository
     {
-        Task<User?> GetUserByIdAsync(int id);
+        Task<User?> GetUserByIdAsync(long id);
         Task<IEnumerable<User>> GetAllUsersAsync();
-        Task<int> CreateUserAsync(User user);
+        Task<long> CreateUserAsync(User user);
     }
 }
